+++ datameta/api/apikeys.py
@@ -18,6 +18,7 @@ from pyramid.request import Request
 from typing import Optional, List
 from .. import models
 from .. import security
+from ..security import authz
 from pyramid.httpexceptions import HTTPOk, HTTPUnauthorized, HTTPForbidden
 from ..resource import resource_by_id, get_identifier
 from ..errors import get_validation_error
@@ -162,7 +163,7 @@ def get_user_keys(request:Request) -> UserSession:
 
     db = request.dbsession
     target_user = resource_by_id(db, models.User, request.matchdict['id'])
-    if not target_user or not security.is_authorized_apikey_view(auth_user, target_user):
+    if not target_user or not authz.is_authorized_apikey_view(auth_user, target_user):
         raise HTTPForbidden()
 
     return ApiKeyLabels(auth_user)
@@ -181,7 +182,7 @@ def delete_key(request:Request) -> UserSession:
 
     db = request.dbsession
     target_key = resource_by_id(db, models.ApiKey, request.matchdict['id'])
-    if not target_key or not security.is_authorized_apikey_deletion(auth_user, target_key.user):
+    if not target_key or not authz.is_authorized_apikey_deletion(auth_user, target_key.user):
         raise HTTPForbidden()
 
     db.delete(target_key)
