+++ datameta/api/password.py
@@ -18,6 +18,7 @@ from sqlalchemy import and_
 from pyramid.view import view_config
 
 from .. import security, errors
+from ..security import authz
 from ..models import User, PasswordToken
 from uuid import UUID
 from datetime import datetime
@@ -64,7 +65,7 @@ def put(request):
             raise HTTPForbidden() # 403 User ID not found, hidden from the user intentionally
 
         # Only changing the user's own password is allowed
-        if not security.is_authorized_password_change(auth_user, target_user) or not security.check_password_by_hash(request_credential, auth_user.pwhash):
+        if not authz.is_authorized_password_change(auth_user, target_user) or not security.check_password_by_hash(request_credential, auth_user.pwhash):
             raise HTTPForbidden() # 403 Not authorized to change this user's password
 
     # Verify the password quality
